Update: Move `PartiallyGraph` in the right place
